fixbug 2366 [夏普] 详情页 包详情页点击更多无响应

+++ appbase/src/main/java/tv/ismar/app/core/PageIntentInterface.java
@@ -52,7 +52,7 @@ public interface PageIntentInterface {
     void toUserCenterLocation(Context context);
 
 
-    void toPackageList(Context context, String source, long pk);
+    void toPackageList(Context context, String source, int pk);
 
     void toFilmStar(Context context, String title, long pk);
